requirements.txt
flask==3.0.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
plotly==5.18.0
geopandas==0.14.1
python-dotenv==1.0.0
requests==2.31.0
shapely==2.0.2
scipy==1.11.4
